Prep for Codgen cleanup

+++ gds/src/core/graph_dimensions.rs
@@ -1,23 +1,23 @@
 /// Graph Dimensions trait - defines the interface for graph dimensions
 pub trait GraphDimensions {
-    fn node_count(&self) -> u64;
-    fn relationship_count(&self) -> u64;
-    fn rel_count_upper_bound(&self) -> u64;
+    fn node_count(&self) -> usize;
+    fn relationship_count(&self) -> usize;
+    fn rel_count_upper_bound(&self) -> usize;
 }
 
 /// Graph Dimensions implementation - placeholder for graph dimensions
 pub struct GraphDimensionsImpl;
 
 impl GraphDimensions for GraphDimensionsImpl {
-    fn node_count(&self) -> u64 {
+    fn node_count(&self) -> usize {
         0
     }
 
-    fn relationship_count(&self) -> u64 {
+    fn relationship_count(&self) -> usize {
         0
     }
 
-    fn rel_count_upper_bound(&self) -> u64 {
+    fn rel_count_upper_bound(&self) -> usize {
         0
     }
 }
@@ -35,34 +35,35 @@ impl Default for GraphDimensionsImpl {
 }
 
 /// Concrete Graph Dimensions - placeholder for concrete graph dimensions
+#[derive(Debug, Clone)]
 pub struct ConcreteGraphDimensions {
-    pub node_count: u64,
+    pub node_count: usize,
     pub relationship_count: usize,
 }
 
 impl GraphDimensions for ConcreteGraphDimensions {
-    fn node_count(&self) -> u64 {
+    fn node_count(&self) -> usize {
         self.node_count
     }
 
-    fn relationship_count(&self) -> u64 {
-        self.relationship_count as u64
+    fn relationship_count(&self) -> usize {
+        self.relationship_count
     }
 
-    fn rel_count_upper_bound(&self) -> u64 {
-        self.relationship_count as u64
+    fn rel_count_upper_bound(&self) -> usize {
+        self.relationship_count
     }
 }
 
 impl ConcreteGraphDimensions {
-    pub fn new(node_count: u64, relationship_count: usize) -> Self {
+    pub fn new(node_count: usize, relationship_count: usize) -> Self {
         Self {
             node_count,
             relationship_count,
         }
     }
 
-    pub fn of(node_count: u64, relationship_count: usize) -> Self {
+    pub fn of(node_count: usize, relationship_count: usize) -> Self {
         Self::new(node_count, relationship_count)
     }
 }
